refactor(solver): move type_references_substitution_keys to CallEvaluator; remove hot-path debug dump
Address reviewer feedback on PR #11866:

1. Move `type_references_substitution_keys` and `substitute_caller_type_params`
   from `call_args.rs` to `call_evaluator.rs` so neither oversized file grows
   above its pre-PR ratchet baseline:
   - `generic_call/resolve.rs`: 3409 → 3380 (below original 3381) ✓
   - `call_args.rs`: 2148 → 2084 (below original 2122) ✓
   - `call_evaluator.rs`: 1466 → 1543 (stays under 2000) ✓

2. Remove the unconditional `tracing::debug!` dump in the per-argument
   constraint loop (hot inference path). Replace the 20-line inline block
   with a one-line call to `self.substitute_caller_type_params(...)`.

3. Update file-size ratchet baselines to the new, smaller values.
