+++ crates/tsz-solver/src/operations/call_args.rs
@@ -1728,70 +1728,6 @@ impl<'a, C: AssignabilityChecker> CallEvaluator<'a, C> {
         }
     }
 
-    /// Recursively check if a type references any `TypeParameter` whose name is a key
-    /// in the given substitution (i.e., one of the caller's type parameter names).
-    pub(crate) fn type_references_substitution_keys(
-        &self,
-        ty: TypeId,
-        substitution: &crate::instantiation::instantiate::TypeSubstitution,
-    ) -> bool {
-        match self.interner.lookup(ty) {
-            Some(TypeData::TypeParameter(info)) => {
-                // Check if this type parameter's name matches one of the caller's type params
-                let name = self.interner.resolve_atom(info.name);
-                if name.as_str().starts_with("__infer_") {
-                    return false; // Inference placeholders are fine
-                }
-                substitution.map().contains_key(&info.name)
-            }
-            Some(TypeData::Union(members)) | Some(TypeData::Intersection(members)) => {
-                let members = self.interner.type_list(members);
-                members
-                    .iter()
-                    .any(|&m| self.type_references_substitution_keys(m, substitution))
-            }
-            Some(TypeData::Array(elem)) => {
-                self.type_references_substitution_keys(elem, substitution)
-            }
-            Some(TypeData::Tuple(elements)) => {
-                let elements = self.interner.tuple_list(elements);
-                elements
-                    .iter()
-                    .any(|e| self.type_references_substitution_keys(e.type_id, substitution))
-            }
-            Some(TypeData::Application(app_id)) => {
-                let app = self.interner.type_application(app_id);
-                app.args
-                    .iter()
-                    .any(|&a| self.type_references_substitution_keys(a, substitution))
-            }
-            // Function signatures extracted from Application types (e.g. from
-            // `Op<A, string>` → `fn(source: A) -> string`) carry caller TypeParameters
-            // in their parameter or return types.
-            Some(TypeData::Function(shape_id)) => {
-                let shape = self.interner.function_shape(shape_id);
-                shape
-                    .params
-                    .iter()
-                    .any(|p| self.type_references_substitution_keys(p.type_id, substitution))
-                    || self.type_references_substitution_keys(shape.return_type, substitution)
-            }
-            Some(TypeData::Callable(shape_id)) => {
-                let shape = self.interner.callable_shape(shape_id);
-                shape
-                    .call_signatures
-                    .iter()
-                    .chain(shape.construct_signatures.iter())
-                    .any(|sig| {
-                        sig.params.iter().any(|p| {
-                            self.type_references_substitution_keys(p.type_id, substitution)
-                        }) || self.type_references_substitution_keys(sig.return_type, substitution)
-                    })
-            }
-            _ => false,
-        }
-    }
-
     #[inline]
     pub(crate) fn type_contains_placeholder(
         &self,
